refactor: move getTransformAttr to utils and move function definations before callling in flowchart converter
