+++ src/constants.ts
@@ -1,2 +1,2 @@
 export const DEFAULT_FONT_SIZE = 20;
-export const SUPPORTED_CHART_TYPES = ["flowchart"];
+export const SUPPORTED_DIAGRAM_TYPES = ["flowchart"];
